Ask user before clearing current configuration
When the user changed the Collector assigned to a Configuration, we
automatically deleted the collector configuration, and tried to set it
to the default template for the Collector (tried but failed, because a bug).

As users may want to keep the current configuration and edit it in the
newly assigned Collector, we now ask users if they want to load the
default template if the configuration input is not empty.